Add hover cursor and visual affordance on clickable graph nodes (#25)
Graph nodes now show cursor:pointer on hover and a subtle glow/size
increase so users can tell they are interactive before clicking.
